refactor(collatz): move pythonkit -> mathkit
