+++ docs/source/conf.py
@@ -54,7 +54,14 @@ def skip_util_classes(app, what, name, obj, skip, options):
     if what == "module":
         skip = True
 
-    modules_to_show = ["optfunckit", "pandaskit", "pythonkit", "sparkkit", "vizkit"]
+    modules_to_show = [
+        "mathkit",
+        "optfunckit",
+        "pandaskit",
+        "pythonkit",
+        "sparkkit",
+        "vizkit",
+    ]
     if what == "module" and any(n in name for n in modules_to_show):
         skip = False
