+++ src/onekit/pythonkit.py
@@ -35,7 +35,6 @@ __all__ = (
     "are_predicates_true",
     "check_random_state",
     "coinflip",
-    "collatz",
     "concat_strings",
     "contrast_sets",
     "create_path",
@@ -252,71 +251,6 @@ def coinflip(bias: float, /, *, seed: Seed = None) -> bool:
     rng = check_random_state(seed)
 
     return rng.random() < bias
-
-
-def collatz(n: int, /) -> Generator:
-    """Generate a Collatz sequence.
-
-    The famous 3n + 1 conjecture [c1]_ [c2]_. Given a positive integer :math:`n > 0`,
-    the next term in the Collatz sequence is half of :math:`n`
-    if :math:`n` is even; otherwise, if :math:`n` is odd,
-    the next term is 3 times :math:`n` plus 1.
-    Symbolically,
-
-    .. math::
-
-        f(n) =
-        \\begin{cases}
-             n / 2 & \\text{ if } n \\equiv 0 \\text{ (mod 2) } \\\\[6pt]
-            3n + 1 & \\text{ if } n \\equiv 1 \\text{ (mod 2) }
-        \\end{cases}
-
-    The Collatz conjecture is that the sequence always reaches 1
-    for any positive integer :math:`n`.
-
-    Parameters
-    ----------
-    n : int
-        A positive integer seeding the Collatz sequence.
-
-    Yields
-    ------
-    int
-        A generator of Collatz numbers that breaks when 1 is reached.
-
-    Raises
-    ------
-    ValueError
-        If ``n`` is not a positive integer.
-
-    References
-    ----------
-    .. [c1] "Collatz", The On-Line Encyclopedia of Integer Sequences®,
-            https://oeis.org/A006370
-    .. [c2] "Collatz conjecture", Wikipedia,
-            https://en.wikipedia.org/wiki/Collatz_conjecture
-
-    Examples
-    --------
-    >>> import toolz
-    >>> import onekit.pythonkit as pk
-    >>> n = 12
-    >>> list(pk.collatz(n))
-    [12, 6, 3, 10, 5, 16, 8, 4, 2, 1]
-    >>> toolz.count(pk.collatz(n))
-    10
-    """
-    if not isinstance(n, int) or n < 1:
-        raise ValueError(f"{n=} - must be a positive integer")
-
-    while True:
-        yield n
-
-        if n == 1:
-            break
-
-        # update
-        n = n // 2 if iseven(n) else 3 * n + 1
 
 
 def concat_strings(sep: str, /, *strings: Iterable[str]) -> str:
